reorg dom range and selection files
